Initial import of code
Some cleanup done.

I suppose this is the point where I'll re-iterate that this is intended
to be a "very simple" program in the nano-style. While I will be
restructuring the code somewhat, it will be primarily so that it will be
easy to bundle this into a mod that is loaded into a larger application,
rather than trying to put down roots and make this more formal via
rebuilding it with Cake or Zend